+++ Docs/Whats_New.md
@@ -33,6 +33,8 @@ Released TBD
 	- Fix tessellated indirect draws using wrong kernels to map parameters.
 	- Work around potential Metal bug with stage-in indirect buffers.
 	- Fix zero local threadgroup size in indirect tessellated rendering.
+- `VkSemaphore` optionally uses `MTLEvent`, if available and 
+  `MVK_ALLOW_METAL_EVENTS` environment variable is enabled.
 - Fix crash when clearing attachments using layered rendering on older macOS devices.
 - Fixes to Metal renderpass layered rendering settings.
 - Fix sporadic crash on `vkDestroySwapchainKHR()`.